课程添加功能

+++ xuecheng-plus-content/xuecheng-plus-content-model/src/main/java/com/xuecheng/content/model/vo/CourseBaseInfoVo.java
@@ -24,12 +24,12 @@ public class CourseBaseInfoVo extends CourseBase {
     /**
      * 价格
      */
-    private BigDecimal price;
+    private Float price;
 
     /**
      * 原价
      */
-    private BigDecimal originalPrice;
+    private Float originalPrice;
 
     /**
      * 咨询qq
